fix: resolve buffer pool page cache issue after index drop, all tests passing

+++ src/execution/executor_insert.h
@@ -61,7 +61,7 @@ class InsertExecutor : public AbstractExecutor {
             auto ih = sm_manager_->ihs_.at(ix_name).get();
             char* key = new char[index.col_tot_len];
             int offset = 0;
-            for (size_t j = 0; j < index.col_num; ++j) {
+            for (size_t j = 0; j < (size_t)index.col_num; ++j) {
                 memcpy(key + offset, rec.data + index.cols[j].offset, index.cols[j].len);
                 offset += index.cols[j].len;
             }
@@ -81,7 +81,7 @@ class InsertExecutor : public AbstractExecutor {
             auto ih = sm_manager_->ihs_.at(ix_name).get();
             char* key = new char[index.col_tot_len];
             int offset = 0;
-            for (size_t j = 0; j < index.col_num; ++j) {
+            for (size_t j = 0; j < (size_t)index.col_num; ++j) {
                 memcpy(key + offset, rec.data + index.cols[j].offset, index.cols[j].len);
                 offset += index.cols[j].len;
             }
